requirements.txt
aiogram==2.22.1
aiohttp==3.8.1
aiosignal==1.2.0
async-timeout==4.0.2
asyncpg==0.26.0
attrs==22.1.0
Babel==2.9.1
certifi==2022.6.15
charset-normalizer==2.1.0
colorama==0.4.5
decorator==4.4.2
frozenlist==1.3.1
idna==3.3
imageio==2.21.1
imageio-ffmpeg==0.4.7
moviepy==1.0.3
multidict==6.0.2
numpy==1.23.2
Pillow==9.2.0
proglog==0.1.10
python-dotenv==0.20.0
pytz==2022.2.1
requests==2.28.1
tqdm==4.64.0
urllib3==1.26.12
yarl==1.8.1
